Task clean_case: Add option to remove additional user-specified files
Accepts a "remove_extra" parameter that must be a list of files and/or
directories that must be removed from the case directory. The entries must be
proper filenames or directories relative to the case directory. No file patterns
are allowed to prevent accidental removal of files the user did not intend to
remove. Might accept file patterns in future versions.